pyproject.toml
[tool.poetry]
name = "api"
version = "0.1"
description = "Sammy API"
authors = ["shavvimal <shavhugan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pytest = "^7.3.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
pydantic = {extras = ["email"], version = "^2.9.2"}
fastapi = "^0.112.4"
uvicorn = "^0.30.6"
gunicorn = "^23.0.0"
langchain-community = "^0.3.1"
langchain-openai = "^0.2.2"
langgraph = "^0.2.34"
tavily-python = "^0.5.0"
tiktoken = "^0.8.0"
langchain = "^0.3.2"
httpx = "^0.26.0"
requests = "^2.32.3"
lxml = "^5.3.0"
aiohttp = "^3.10.10"
twilio = "^9.3.4"
supabase = "^2.9.1"
cryptography = "^43.0.3"
asyncpg = "^0.30.0"
python-multipart = "^0.0.17"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
boto3 = "^1.35.44"

[build-system]
# A list of dependencies required to build your package, making this key mandatory
requires = ["poetry-core"]
# The Python object used to perform the build process
build-backend = "poetry.core.masonry.api"